requirements.txt
Django==1.8
django-ajax-selects==1.4.0
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.6.0
django-messages==0.5.3
django-social-auth==0.7.28
django-tables2==1.4.1
django-taggit==0.17.1
djangotoolbox==1.8.0
psycopg2==2.6.2
xhtml2pdf==0.0.6
xlrd==1.0.0
zope.interface==4.1.2
whitenoise==3.2
#uwsgi==2.0.15
html5lib==1.b8
xlwt==1.2.0
openpyxl==2.4.7
